Se agrega ultimos cambios para confirmacion despacho, programacion arribo

+++ portalBachuco/infrastructure/src/main/java/com/bachoco/persistence/adapter/SiloJpaAdapter.java
@@ -3,8 +3,10 @@ package com.bachoco.persistence.adapter;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 
+import com.bachoco.exception.CannotDeleteResourceException;
 import com.bachoco.mapper.SiloMapper;
 import com.bachoco.model.Silo;
 import com.bachoco.persistence.entity.SiloEntity;
@@ -12,34 +14,34 @@ import com.bachoco.persistence.repository.SiloJpaRepository;
 import com.bachoco.port.SiloRepositoryPort;
 
 @Component
-public class SiloJpaAdapter implements SiloRepositoryPort{
-	
+public class SiloJpaAdapter implements SiloRepositoryPort {
+
 	private final SiloJpaRepository siloJpaRepository;
-	
+
 	public SiloJpaAdapter(SiloJpaRepository siloJpaRepository) {
 		this.siloJpaRepository = siloJpaRepository;
 	}
 
 	@Override
-	public 	Optional<Silo> save(Silo model) {
-		SiloEntity entity=this.siloJpaRepository.save(SiloMapper.toEntitySave(model));
+	public Optional<Silo> save(Silo model) {
+		SiloEntity entity = this.siloJpaRepository.save(SiloMapper.toEntitySave(model));
 		return Optional.ofNullable(SiloMapper.toDomain(entity));
 	}
 
 	@Override
 	public Optional<Silo> update(Silo model) {
 		try {
-			Optional<SiloEntity> siloEntity= this.siloJpaRepository.findById(model.getId());
-			if(siloEntity.isPresent()) {
-				SiloEntity siloUpdate=siloEntity.get();
+			Optional<SiloEntity> siloEntity = this.siloJpaRepository.findById(model.getId());
+			if (siloEntity.isPresent()) {
+				SiloEntity siloUpdate = siloEntity.get();
 				siloUpdate.setSilo(model.getSilo());
 				siloUpdate.setNombre(model.getNombre());
 				siloUpdate.setSociedad(model.getSociedad());
-				SiloEntity updated= this.siloJpaRepository.save(siloUpdate);
-				return Optional.ofNullable(SiloMapper.toDomain(updated)); 
+				SiloEntity updated = this.siloJpaRepository.save(siloUpdate);
+				return Optional.ofNullable(SiloMapper.toDomain(updated));
 			}
 			return Optional.empty();
-		}catch (Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
 		}
 		return Optional.empty();
@@ -47,11 +49,14 @@ public class SiloJpaAdapter implements SiloRepositoryPort{
 
 	@Override
 	public void delete(Integer id) {
-		try {
-			Optional<SiloEntity> siloEntity= this.siloJpaRepository.findById(id);
-			this.siloJpaRepository.delete(siloEntity.get());
-		}catch (Exception e) {
-			// TODO: handle exception
+		Optional<SiloEntity> siloEntity = this.siloJpaRepository.findById(id);
+		if (siloEntity.isPresent()) {
+			try {
+				this.siloJpaRepository.delete(siloEntity.get());
+			} catch (DataIntegrityViolationException e) {
+				throw new CannotDeleteResourceException("No se puede eliminar el producto, existen pedidos asociados.",
+						e);
+			}
 		}
 	}
 
@@ -60,15 +65,14 @@ public class SiloJpaAdapter implements SiloRepositoryPort{
 		return SiloMapper.toListDomain(this.siloJpaRepository.findAll());
 	}
 
-	
 	@Override
 	public Optional<Silo> findById(Integer id) {
 		try {
-			Optional<SiloEntity> siloEntity= this.siloJpaRepository.findById(id);
-			if(siloEntity.isPresent()) {
+			Optional<SiloEntity> siloEntity = this.siloJpaRepository.findById(id);
+			if (siloEntity.isPresent()) {
 				return Optional.ofNullable(SiloMapper.toDomain(siloEntity.get()));
 			}
-		}catch (Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
 		}
 		return Optional.empty();
@@ -77,11 +81,11 @@ public class SiloJpaAdapter implements SiloRepositoryPort{
 	@Override
 	public Optional<Float> stock(Integer siloId) {
 		try {
-			Optional<Float> siloEntity= this.siloJpaRepository.findStockById(siloId);
-			if(siloEntity.isPresent()) {
+			Optional<Float> siloEntity = this.siloJpaRepository.findStockById(siloId);
+			if (siloEntity.isPresent()) {
 				return Optional.ofNullable(siloEntity.get());
 			}
-		}catch (Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
 		}
 		return Optional.empty();
@@ -89,8 +93,8 @@ public class SiloJpaAdapter implements SiloRepositoryPort{
 
 	@Override
 	public Optional<Silo> findByClave(String clave) {
-		Optional<SiloEntity> entity=this.siloJpaRepository.findBySilo(clave);
-		if(entity.isPresent()) {
+		Optional<SiloEntity> entity = this.siloJpaRepository.findBySilo(clave);
+		if (entity.isPresent()) {
 			return Optional.ofNullable(SiloMapper.toDomain(entity.get()));
 		}
 		return Optional.empty();
